PRC-52: Cleanup
Renamed classes into consistent plurality and improved the naming of the integration tests. Split the contact + prisoner contact services. Moved the model into request and response packages. Renamed the entities for prisoner contact summary to be consistent by using entity suffix.
