+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/repository/PrisonerContactRepository.kt
@@ -2,17 +2,17 @@ package uk.gov.justice.digital.hmpps.hmppscontactsapi.repository
 
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.entity.PrisonerContactDetail
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.entity.PrisonerContactSummaryEntity
 
 @Repository
-interface PrisonerContactRepository : ReadOnlyRepository<PrisonerContactDetail, Long> {
+interface PrisonerContactRepository : ReadOnlyRepository<PrisonerContactSummaryEntity, Long> {
 
   @Query(
     value = """
-      FROM PrisonerContactDetail pcs
+      FROM PrisonerContactSummaryEntity pcs
       WHERE pcs.prisonerNumber = :prisonerNumber
       AND pcs.active = :activeFlag
     """,
   )
-  fun findPrisonerContacts(prisonerNumber: String, activeFlag: Boolean = true): List<PrisonerContactDetail>
+  fun findPrisonerContacts(prisonerNumber: String, activeFlag: Boolean = true): List<PrisonerContactSummaryEntity>
 }
