--- src/test/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/integration/resource/ContactEntityControllerResourceIntegrationTest.kt
@@ -1,130 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppscontactsapi.integration.resource
-
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.http.MediaType
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.entity.ContactEntity
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.integration.IntegrationTestBase
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.CreateContactRequest
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.repository.ContactRepository
-import java.time.LocalDate
-
-class ContactEntityControllerResourceIntegrationTest : IntegrationTestBase() {
-
-  @Autowired
-  private lateinit var contactRepository: ContactRepository
-
-  @Nested
-  inner class Create {
-    @Test
-    fun `should return unauthorized if no token`() {
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(aMinimalCreateContactRequest())
-        .exchange()
-        .expectStatus()
-        .isUnauthorized
-    }
-
-    @Test
-    fun `should return forbidden if no role`() {
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(aMinimalCreateContactRequest())
-        .headers(setAuthorisation())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return forbidden if wrong role`() {
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(aMinimalCreateContactRequest())
-        .headers(setAuthorisation(roles = listOf("ROLE_WRONG")))
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return bad request if required fields are null`() {
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .headers(setAuthorisation(roles = listOf("ROLE_CONTACTS_ADMIN")))
-        .bodyValue("{}")
-        .exchange()
-        .expectStatus()
-        .isBadRequest
-    }
-
-    @Test
-    fun `should create the contact with minimal fields`() {
-      val request = aMinimalCreateContactRequest()
-
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .headers(setAuthorisation(roles = listOf("ROLE_CONTACTS_ADMIN")))
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isCreated
-
-      assertThatContactIsCreated(request)
-    }
-
-    @Test
-    fun `should create the contact with all fields`() {
-      val request = CreateContactRequest(
-        title = "mr",
-        lastName = "last",
-        firstName = "first",
-        middleName = "middle",
-        dateOfBirth = LocalDate.of(1982, 6, 15),
-        createdBy = "created",
-      )
-
-      webTestClient.post()
-        .uri("/contact")
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .headers(setAuthorisation(roles = listOf("ROLE_CONTACTS_ADMIN")))
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isCreated
-
-      assertThatContactIsCreated(request)
-    }
-
-    private fun assertThatContactIsCreated(request: CreateContactRequest) {
-      with(contactRepository.findAll().maxBy(ContactEntity::contactId)) {
-        assertThat(title).isEqualTo(request.title)
-        assertThat(lastName).isEqualTo(request.lastName)
-        assertThat(firstName).isEqualTo(request.firstName)
-        assertThat(middleName).isEqualTo(request.middleName)
-        assertThat(dateOfBirth).isEqualTo(request.dateOfBirth)
-        assertThat(createdBy).isEqualTo(request.createdBy)
-      }
-    }
-
-    private fun aMinimalCreateContactRequest() = CreateContactRequest(
-      lastName = "last",
-      firstName = "first",
-      createdBy = "created",
-    )
-  }
-}
